Fix logging of error with no message
Today my app received an error with no message, and so toolbelt failed when trying to log it, saying: `TypeError: Cannot read property 'replace' of undefined`.

It's bad to have errors without messages, but I've added a check so that toolbelt at least won't fail when that happens, and we can ask for more error details in these cases as we detect them.