Fix error when installing an app provided as command argument
As we were always trying to read the manifest even when
an app was provided in the command line, the install command
could only be called from inside an app folder. This is not
necessary (nor correct) as the provided app ID is enough to
know which app to (try to) install.