Check subject inside messages
When publishing or linking an app that is not available in the dependency tree, the events sent by builder hub are crashing in colossus, because it checks if the subject - the app you are linking or publishing - is in dependency tree, which is not the case.

Now builder hub will send the message to any subject (`-`), and add the right subject inside the message.