Fix getting environment from process env
To be less confusing, allow the process env also
to set the env to production, by only ignoring it
if it is actually unset.

The other way around, if the user had a configuration
for the staging environment and an env var VTEX_ENV=prod
(or anything different than 'beta'), it'd keep using the
configuration instead of deriving the non-beta environment
from the env var.

This now only falls back to the configuration if the env
var is really not set.