Allow forcing production IO env via process env
We were only considering the VTEX_ENV variable when it
was set to 'beta' to force toolbelt to use the staging env.

A more useful case tho is to force toolbelt to use
production tho, by considering when the env is set to
'prod' instead. This will allow for VTEX users to keep using
prod if they have some demand for it (case for developing
checkout extensions or GoCommerce apps).