Fix yarn.lock for ramda package
I do not know how this happened or fixed itself, but
I was getting some weird type errors from ramda on `yarn`,
then I decided to remove the `@types/ramda` dependency from
the `package.json`. That made the build work and `yarn`
apparently updated the `yarn.lock` for the ramda package.

Then adding the `@types/ramda` back to `package.json` the
same way it was before, keeps working, so I'm guessing that
change to the `yarn.lock` is what fixed it (when I remove
that change the build starts failing again...)