Fix check before setting a workspace weight
Unlike what was being done, we should not simply skip the production command.
Instead, a workspace must already be set to production before we can set its
weight and we should not change the value of that flag.

Checking for links instead of checking the production flag is also bad because
it encourages a bad behavior by the users, which is promoting a workspace that
has had links. This is going to be forbidden in the future, so we should always
nudge the user in the right direction with the error messages.