Some fixes to workspace handling for production and AB test (#477)
* Change error message from `vtex production true` when links exist

We do not want people to promote workspaces that have had links before. While we
do not enforce the correct workspace lifecycle, at least instruct the users to
do the correct thing when setting stuff to production, which is reset said
workspace or create a new one.

* Get the old behavior of `vtex production` back

People used to run `vtex production` and it would set the argument to true.
Today I have received reports that this is not happened and it appears to be
because the value of the argument is undefined, not null. To fix that, accept
both possibilities when handling the argument.

* Fix check before setting a workspace weight

Unlike what was being done, we should not simply skip the production command.
Instead, a workspace must already be set to production before we can set its
weight and we should not change the value of that flag.

Checking for links instead of checking the production flag is also bad because
it encourages a bad behavior by the users, which is promoting a workspace that
has had links. This is going to be forbidden in the future, so we should always
nudge the user in the right direction with the error messages.

* Update A/B test error message