Implement most of the states and transitions as chronos expects (#509)
We do not allow `dev` workspaces to become `production` workspaces and the
opposite is not something we should do either (`dev` and `production` are mostly
separate worlds).

This means that a workspace reset (re-create the workspace
from scratch) should be allowed to create the workspace as any of those, but
`vtex production` no longer exists, neither with `true` nor with `false` as the
last argument. To phase out the command, first mark it as deprecated in the list
shown by `vtex help` and replace the actual command with an error message with
instructions.

Another consequence is that links are not allowed in a production workspace.
This change should avoid that for most cases, though we still need to update
apps to really forbid it for good.

I had also to update a few messages that instructed the user to change the
workspace state to production. All these messages should now instruct the user
to either create or reset a workspace to make it a production one.

Finally, I felt that the information about the workspace state was not so
evident for a user, so I have created a new command to allow one to answer the
question "is the workspace I am using a production or a dev workspace?" It is
called `status` because it was the best word I could find for it, but I am not
100% happy with it. My feeling is that `production` was the best name, but it is
already taken, so I would gladly accept suggestions of better names for this new
command.

As far as I can tell, I have changed all the places that we have to change now.
We will still need to adapt Kube Router (and the whole platform) to the new
"beta" concept that came up this week, but that is for later.