Stop making `vtex update` attempt to update app major
Trying to update the major makes the API call to apps fail, as the
apps service currently doesn't allow an installation request to change
the major of an existing installation. I'm removing that check in
apps, but for that we need to make toolbelt not accidentally update
the apps majors in users' accounts.

That would erroneously make many people update their apps to
the latest majors, when many times the latest major are actually
"unstable" and kept in active development until they are actually
released (when we would normally update the accounts automatically).
This is currently the case for many apps such as vtex.admin, vtex.store,
render-server, render-ssr, etc.