Move functionality from conflicting command
`vtex browse admin` was an awesome idea that got incorporated into toolbelt
before we got to merge our code. The master branch now has a more general
version of the command (`vtex browse`) and it conflicted with our
implementation. The solution is to incorporate our support functionality to the
code that is already in master and get rid of our own command.