Use HTTPS when handling editions
The addition to `clientOptions` is meant to make cluster selection seamless
wherever the constant is used. The `cluster()` value is allowed to be anything,
as we only take it into account if it is a test cluster. This means that nothing
will break if you run `vtex config set cluster banana` and try to run a command.
The command will just happen in whatever production cluster should get it
anyway.