Set the edition in the correct workspace
As we are calling the ensureValidEdition from the
create and reset commands now, it is not necessarily
the case that we will already be in that specific
workspace when setting the edition through that
validation logic.

For that, created an optional argument to the setEdition
command, receiving the workspace that we should set the
edition on. Also created an option to automatically switch
to the previous configuration, as the ensureValidEdition
might be running in the middle of a `vtex use` command,
and we don't want to risk giving the user the option of
ending up with an invalid configuration.
e.g. If they'd answer not to return to the previous account,
we would stay in the sponsor account and then the vtex use
command would finish "using" the new workspace, changing only
the workspace in the configuration. That would end up with
an invalid configuration, as the user was originally creating
a workspace in the child account, not in the sponsor and the
confusion could stack up a lot.